Extract out common bits from Dttp:Retrieve services
